tests: Move Citrea deposit test to deposit.rs.
